requirements.txt
arrow==0.13.1
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
Django==2.0.1
django-webpush==0.2.7
http-ece==1.1.0
idna==2.8
Jinja2==2.10.1
MarkupSafe==1.1.1
py-vapid==1.5.0
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.1
pywebpush==1.6.0
requests==2.21.0
six==1.12.0
urllib3==1.24.2
